Refactor test for all add on options
